series

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -2,16 +2,16 @@
 
 int main(void)
 {
-long long int a = 0;
-long long int s = 1;
-long long int d;
+unsigned long int a = 0;
+unsigned long int s = 1;
+unsigned long int d;
 int f;
 for (f = 1; f < 98; f++)
 	{
 	d = a + s; 
 	a = s; 
 	s = d; 
-	printf("%lld, ", s);
+	printf("%lu, ", s);
 	}
 printf("\n");
 return (0);
